# swagger.ru.yaml
openapi: 3.0.0
info:
  title: TigerMeter Cloud API
  description: API для привязки (claim) устройств и управления универсальными инструкциями отображения с настраиваемыми текстовыми полями (symbol, topLine, mainText, bottomLine), размерами шрифтов, выравниванием и LED-управлением.
  version: 2.0.0
  x-language: ru
servers:
  - url: https://api.tigermeter.cloud
    description: Prod
  - url: http://localhost:3000
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/devices:
    get:
      tags: [Portal]
      summary: Список моих устройств
      operationId: listMyDevices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Устройства
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceSummary'
        '401': { description: Не авторизован }
  /api/devices/{id}:
    get:
      tags: [Portal]
      summary: Детали устройства
      operationId: getDevice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Подробности устройства
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetail'
        '401': { description: Не авторизован }
        '403': { description: Запрещено }
        '404': { description: Не найдено }
  /api/devices/{id}/revoke:
    post:
      tags: [Portal]
      summary: Отвязать (revoke) устройство
      operationId: revokeDevice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Устройство отвязано
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: revoked }
        '401': { description: Не авторизован }
        '403': { description: Запрещено }
        '404': { description: Не найдено }
  /api/devices/{id}/display:
    put:
      tags: [Portal]
      summary: Установить инструкцию отображения
      operationId: setDisplayInstruction
      description: Заменяет текущую инструкцию отображения. Устройство должно переключиться при изменении хеша.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisplayInstruction'
            examples:
              example:
                $ref: '#/components/examples/InstructionExample'
      responses:
        '200':
          description: Принято; возвращает текущий хеш
          content:
            application/json:
              schema:
                type: object
                properties:
                  displayHash: { type: string }
        '400': { description: Неверный запрос }
        '401': { description: Не авторизован }
        '403': { description: Запрещено }
        '404': { description: Не найдено }
  /api/devices/{id}/heartbeat:
    post:
      tags: [Device]
      summary: Хартбит
      description: Возвращает ok или новую инструкцию отображения если она изменилась.
      operationId: heartbeat
      security:
        - deviceSecretAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceHeartbeat'
      responses:
        '200':
          description: Результат хартбита
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/HeartbeatOk'
                  - $ref: '#/components/schemas/HeartbeatWithInstruction'
        '401': { description: Недействительный секрет }
        '404': { description: Не найдено }
  /api/devices/{id}/display/hash:
    get:
      tags: [Device]
      summary: Получить текущий хеш инструкции
      operationId: getDisplayHash
      security:
        - deviceSecretAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Текущий хеш
          content:
            application/json:
              schema:
                type: object
                properties:
                  hash: { type: string, description: Хеш текущей инструкции }
        '401': { description: Недействительный секрет }
        '404': { description: Не найдено }
  /api/devices/{id}/display/full:
    get:
      tags: [Device]
      summary: Получить полную инструкцию отображения
      operationId: getDisplayInstruction
      security:
        - deviceSecretAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: ifHash
          in: query
          schema:
            type: string
          description: Если совпадает с текущим хешем - 304
      responses:
        '200':
          description: Инструкция отображения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisplayInstruction'
              examples:
                instruction:
                  $ref: '#/components/examples/InstructionExample'
        '304': { description: Не изменилось }
        '401': { description: Недействительный секрет }
        '404': { description: Не найдено }
  /api/devices/{id}/secret/refresh:
    post:
      tags: [Device]
      summary: Обновить секрет устройства
      operationId: refreshSecret
      security:
        - deviceSecretAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Новый секрет выдан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimPollClaimed'
        '401': { description: Недействительный секрет }
        '404': { description: Не найдено }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceSecretAuth:
      type: http
      scheme: bearer
      bearerFormat: Device Secret
      description: Секретный токен устройства
  schemas:
    DisplayInstruction:
      type: object
      required: [version, hash, symbol, mainText]
      description: Инструкция отображения (плоская структура).
      properties:
        version:
          type: integer
          description: Монотонно растущая версия инструкции для устройства
        hash:
          type: string
          description: Стабильный хеш JSON инструкции
        symbol: { type: string, description: Короткий идентификатор (тикер) }
        mainText: { type: string, description: Основной текст по центру }
        topLine: { type: string, description: Текст верхней строки }
        topLineFontSize: { type: integer, minimum: 10, maximum: 40, default: 16, description: 'Размер шрифта верхней строки (10-40px, пресеты: 10, 12, 16, 20, 24, 28, 32, 40)' }
        topLineAlign: { type: string, enum: [left, center, right], default: center, description: Выравнивание верхней строки }
        topLineShowDate: { type: boolean, default: false, description: Показывать текущее время вместо topLine }
        mainTextFontSize: { type: integer, minimum: 10, maximum: 40, default: 32, description: 'Размер шрифта основного текста (10-40px, пресеты: 10, 12, 16, 20, 24, 28, 32, 40)' }
        mainTextAlign: { type: string, enum: [left, center, right], default: center, description: Выравнивание основного текста }
        bottomLine: { type: string, description: Текст нижней строки }
        bottomLineFontSize: { type: integer, minimum: 10, maximum: 40, default: 16, description: 'Размер шрифта нижней строки (10-40px, пресеты: 10, 12, 16, 20, 24, 28, 32, 40)' }
        bottomLineAlign: { type: string, enum: [left, center, right], default: center, description: Выравнивание нижней строки }
        ledColor: { type: string, enum: [blue, green, red, yellow, purple], default: green, description: Цвет светодиода }
        ledBrightness: { type: string, enum: [off, low, mid, high], default: mid, description: Яркость LED }
        beep: { type: boolean, default: false, description: Однократный звуковой сигнал (one-time) }
        flashCount: { type: integer, minimum: 0, default: 0, description: Количество вспышек LED (one-time) }
        refreshInterval: { type: integer, minimum: 5, default: 30, description: Интервал опроса API в секундах }
        extensions:
          type: object
          additionalProperties: true
          description: Зарезервировано для будущих расширений
    DeviceSummary:
      type: object
      description: Краткая информация об устройстве.
      properties:
        id: { type: string, description: Идентификатор устройства }
        name: { type: string, description: Пользовательское имя }
        status: { type: string, enum: [awaiting_claim, active, revoked], description: Статус устройства }
        lastSeen: { type: string, format: date-time, nullable: true, description: Последний хартбит }
    DeviceDetail:
      allOf:
        - $ref: '#/components/schemas/DeviceSummary'
        - type: object
          description: Детализированная информация устройства.
          properties:
            mac: { type: string, description: MAC адрес }
            userId: { type: string, nullable: true, description: Пользователь-владелец }
            battery: { type: integer, nullable: true, description: Заряд батареи (проценты) }
            led: { type: string, enum: [off, low, middle, high], nullable: true, description: Текущая яркость LED }
            secretExpiresAt: { type: string, format: date-time, nullable: true, description: Истечение текущего секрета }
            displayHash: { type: string, description: Текущий хеш инструкции }
...
    DeviceHeartbeat:
      type: object
      description: Данные хартбита устройства.
      properties:
        battery: { type: integer, description: Заряд батареи }
        rssi: { type: integer, description: Уровень сигнала }
        ip: { type: string, description: IP адрес устройства }
        firmwareVersion: { type: string, description: Версия прошивки }
        uptimeSeconds: { type: integer, description: Аптайм в секундах }
        displayHash: { type: string, description: Хеш который устройство считает актуальным }
    HeartbeatOk:
      type: object
      description: Ответ когда изменений нет.
      properties:
        ok:
          type: boolean
          example: true
          description: Инструкций к обновлению нет
    HeartbeatWithInstruction:
      type: object
      description: Ответ с новой инструкцией отображения.
      properties:
        instruction:
          $ref: '#/components/schemas/DisplayInstruction'
        displayHash: { type: string, description: Новый хеш инструкции }
...
    ClaimPollClaimed:
      type: object
      description: Одноразовый ответ при успешной привязке. Содержит секрет устройства, который генерируется лениво в момент первого успешного опроса. Повторные опросы после получения секрета возвращают 404.
      properties:
        deviceId: { type: string, description: Идентификатор устройства }
        deviceSecret: { type: string, description: Новый секрет устройства (plaintext; показывается только один раз) }
        displayHash: { type: string, description: Текущий хеш инструкции (может быть пустым) }
        expiresAt: { type: string, format: date-time, description: Истечение секрета }
  examples:
    InstructionExample:
      summary: Инструкция отображения
      value:
        version: 7
        hash: "sha256:ab12cd34ef56..."
        symbol: BTC
        mainText: "$98,500"
        topLine: Bitcoin
        topLineFontSize: 16
        topLineAlign: center
        mainTextFontSize: 32
        mainTextAlign: center
        bottomLine: "+2.5%"
        bottomLineFontSize: 16
        bottomLineAlign: center
        ledColor: green
        ledBrightness: mid
        beep: false
        flashCount: 0
        refreshInterval: 30
...